Update readme.md
"updatePrice" call should not be present in the first part of the code where we are just add/removing items. This throws a method not found exception in the console window. 

In the second part, the prefix of "model.controller" should only be on the "onAtbClick" method (which is the button click handler). Otherwise, we get the same error and clicking the "add to basket" doesn't make any price changes.